fix(*): cleanups
